Reorganize core unit tests, Add variable documentation/tests
