+++ tests/core/test_collections.py
@@ -1,0 +1,275 @@
+"""Unit tests for core.collections module."""
+import unittest
+
+import numpy as np
+
+import relentless
+
+class test_PairMatrix(unittest.TestCase):
+    """Unit tests for relentless.PairMatrix."""
+
+    def test_init(self):
+        """Test construction with different list types."""
+        types = ('A','B')
+        pairs  = (('A','B'), ('B','B'), ('A','A'))
+
+        #test construction with tuple input
+        m = relentless.PairMatrix(types=('A','B'))
+        self.assertEqual(m.types, types)
+        self.assertCountEqual(m.pairs, pairs)
+
+        #test construction with list input
+        m = relentless.PairMatrix(types=['A','B'])
+        self.assertEqual(m.types, types)
+        self.assertCountEqual(m.pairs, pairs)
+
+        types = ('A',)
+        pairs = (('A','A'),)
+
+        #test construction with single type tuple
+        m = relentless.PairMatrix(types=('A',))
+        self.assertEqual(m.types, types)
+        self.assertCountEqual(m.pairs, pairs)
+
+        #test construction with int type input
+        with self.assertRaises(TypeError):
+            m = relentless.PairMatrix(types=(1,2))
+
+        #test construction with mixed type input
+        with self.assertRaises(TypeError):
+            m = relentless.PairMatrix(types=('1',2))
+
+    def test_accessors(self):
+        """Test get and set methods on pairs."""
+        m = relentless.PairMatrix(types=('A','B'))
+
+        #test set and get for each pair type
+        m['A','A']['energy'] = 1.0
+        self.assertEqual(m['A','A']['energy'], 1.0)
+        self.assertEqual(m['A','B'], {})
+        self.assertEqual(m['B','B'], {})
+
+        m['A','B']['energy'] = -1.0
+        self.assertEqual(m['A','A']['energy'], 1.0)
+        self.assertEqual(m['A','B']['energy'], -1.0)
+        self.assertEqual(m['B','B'], {})
+
+        m['B','B']['energy'] = 1.0
+        self.assertEqual(m['A','A']['energy'], 1.0)
+        self.assertEqual(m['A','B']['energy'], -1.0)
+        self.assertEqual(m['B','B']['energy'], 1.0)
+
+        #test key order equality
+        self.assertEqual(m['A','B'], m['B','A'])
+
+        #test re-set and get
+        m['A','A']['energy'] = 2.0
+        self.assertEqual(m['A','A']['energy'], 2.0)
+        self.assertEqual(m['A','B']['energy'], -1.0)
+        self.assertEqual(m['B','B']['energy'], 1.0)
+
+        m['A','B']['energy'] = -1.5
+        self.assertEqual(m['A','A']['energy'], 2.0)
+        self.assertEqual(m['A','B']['energy'], -1.5)
+        self.assertEqual(m['B','B']['energy'], 1.0)
+
+        m['B','B']['energy'] = 0.0
+        self.assertEqual(m['A','A']['energy'], 2.0)
+        self.assertEqual(m['A','B']['energy'], -1.5)
+        self.assertEqual(m['B','B']['energy'], 0.0)
+
+        #test setting multiple parameters and get
+        m['A','A']['mass'] = 1.0
+        self.assertEqual(m['A','A']['mass'], 1.0)
+        self.assertEqual(m['A','A']['energy'], 2.0)
+        self.assertEqual(m['A','A'], {'energy':2.0, 'mass':1.0})
+
+        m['A','B']['mass'] = 3.0
+        self.assertEqual(m['A','B']['mass'], 3.0)
+        self.assertEqual(m['A','B']['energy'], -1.5)
+        self.assertEqual(m['A','B'], {'energy':-1.5, 'mass':3.0})
+
+        m['B','B']['mass'] = 5.0
+        self.assertEqual(m['B','B']['mass'], 5.0)
+        self.assertEqual(m['B','B']['energy'], 0.0)
+        self.assertEqual(m['B','B'], {'energy':0.0, 'mass':5.0})
+
+        #test setting paramters for invalid keys
+        with self.assertRaises(KeyError):
+            x = m['C','C']
+        with self.assertRaises(KeyError):
+            x = m['A','C']
+
+    def test_iteration(self):
+        """Test iteration on the matrix."""
+        m = relentless.PairMatrix(types=('A','B'))
+
+        #test iteration for initialization
+        for pair in m:
+            m[pair]['mass'] = 2.0
+            m[pair]['energy'] = 1.0
+        self.assertEqual(m['A','B'], {'energy':1.0, 'mass':2.0})
+        self.assertEqual(m['A','A'], {'energy':1.0, 'mass':2.0})
+        self.assertEqual(m['B','B'], {'energy':1.0, 'mass':2.0})
+
+        #test resetting values manually
+        m['A','B']['mass'] = 2.5
+        m['A','A']['energy'] = 1.5
+        self.assertEqual(m['A','B'], {'energy':1.0, 'mass':2.5})
+        self.assertEqual(m['A','A'], {'energy':1.5, 'mass':2.0})
+        self.assertEqual(m['B','B'], {'energy':1.0, 'mass':2.0})
+
+        #test re-iteration for setting values
+        for pair in m:
+            m[pair]['energy'] = 3.0
+        self.assertEqual(m['A','B'], {'energy':3.0, 'mass':2.5})
+        self.assertEqual(m['A','A'], {'energy':3.0, 'mass':2.0})
+        self.assertEqual(m['B','B'], {'energy':3.0, 'mass':2.0})
+
+class test_FixedKeyDict(unittest.TestCase):
+    """Unit tests for relentless.FixedKeyDict."""
+
+    def test_init(self):
+        """Test construction with different list keys."""
+        keys = ('A','B')
+        default = {'A':1.0, 'B':1.0}
+
+        #test construction with tuple input
+        d = relentless.FixedKeyDict(keys=('A','B'))
+        self.assertEqual(d.keys, keys)
+        self.assertEqual([d[k] for k in d.keys], [None, None])
+
+        #test construction with list input
+        d = relentless.FixedKeyDict(keys=['A','B'])
+        self.assertEqual(d.keys, keys)
+        self.assertEqual([d[k] for k in d.keys], [None, None])
+
+        #test construction with defined default input
+        d = relentless.FixedKeyDict(keys=('A','B'), default=1.0)
+        self.assertEqual(d.keys, keys)
+        self.assertEqual([d[k] for k in d.keys], [1.0, 1.0])
+
+        #test construction with single-key tuple input
+        keys = ('A',)
+        d = relentless.FixedKeyDict(keys=('A',))
+        self.assertEqual(d.keys, keys)
+        self.assertEqual([d[k] for k in d.keys], [None])
+
+        #test construction with int key input
+        with self.assertRaises(TypeError):
+            d = relentless.FixedKeyDict(keys=(1,2))
+
+        #test construction with mixed key input
+        with self.assertRaises(TypeError):
+            d = relentless.FixedKeyDict(keys=('1',2))
+
+    def test_accessors(self):
+        """Test get and set methods on keys."""
+        d = relentless.FixedKeyDict(keys=('A','B'))
+
+        #test setting and getting values
+        d['A'] = 1.0
+        self.assertEqual([d[k] for k in d.keys], [1.0, None])
+        d['B'] = 1.0
+        self.assertEqual([d[k] for k in d.keys], [1.0, 1.0])
+
+        #test re-setting and getting values
+        d['A'] = 2.0
+        self.assertEqual([d[k] for k in d.keys], [2.0, 1.0])
+        d['B'] = 1.5
+        self.assertEqual([d[k] for k in d.keys], [2.0, 1.5])
+
+        #test getting invalid key
+        with self.assertRaises(KeyError):
+            x = d['C']
+
+    def test_update(self):
+        """Test update method to get and set keys."""
+        d = relentless.FixedKeyDict(keys=('A','B'))
+
+        self.assertEqual([d[k] for k in d.keys], [None, None])
+
+        #test updating both keys
+        d.update({'A':1.0, 'B':2.0})  #using dict
+        self.assertEqual([d[k] for k in d.keys], [1.0, 2.0])
+
+        d.update(A=1.5, B=2.5)  #using kwargs
+        self.assertEqual([d[k] for k in d.keys], [1.5, 2.5])
+
+        #test updating only one key at a time
+        d.update({'A':1.1})   #using dict
+        self.assertEqual([d[k] for k in d.keys], [1.1, 2.5])
+
+        d.update(B=2.2)   #using kwargs
+        self.assertEqual([d[k] for k in d.keys], [1.1, 2.2])
+
+        #test using *args length > 1
+        with self.assertRaises(TypeError):
+            d.update({'A':3.0}, {'B':4.0})
+
+        #test using both *args and **kwargs
+        d.update({'A':3.0, 'B':2.0}, B=2.2)
+        self.assertEqual([d[k] for k in d.keys], [3.0, 2.2])
+
+        #test using invalid kwarg
+        with self.assertRaises(KeyError):
+            d.update(C=2.5)
+
+    def test_clear(self):
+        """Test clear method to reset keys to default."""
+        #test clear with no default set
+        d = relentless.FixedKeyDict(keys=('A','B'))
+        self.assertEqual([d[k] for k in d.keys], [None, None])
+        d.update(A=2, B=3)
+        self.assertEqual([d[k] for k in d.keys], [2.0, 3.0])
+        d.clear()
+        self.assertEqual([d[k] for k in d.keys], [None, None])
+
+        #test clear with set default
+        d = relentless.FixedKeyDict(keys=('A','B'), default=1.0)
+        self.assertEqual([d[k] for k in d.keys], [1.0, 1.0])
+        d.update(A=2, B=3)
+        self.assertEqual([d[k] for k in d.keys], [2.0, 3.0])
+        d.clear()
+        self.assertEqual([d[k] for k in d.keys], [1.0, 1.0])
+
+    def test_iteration(self):
+        """Test iteration on the dictionary."""
+        d = relentless.FixedKeyDict(keys=('A','B'))
+
+        #test iteration for setting values
+        for k in d:
+            d[k] = 1.0
+        self.assertEqual([d[k] for k in d.keys], [1.0, 1.0])
+
+        #test manual re-setting of values
+        d['A'] = 2.0
+        self.assertEqual([d[k] for k in d.keys], [2.0, 1.0])
+        d['B'] = 1.5
+        self.assertEqual([d[k] for k in d.keys], [2.0, 1.5])
+
+        #test iteration for re-setting values
+        for k in d:
+            d[k] = 3.0
+        self.assertEqual([d[k] for k in d.keys], [3.0, 3.0])
+
+    def test_copy(self):
+        """Test copying custom dict to standard dict."""
+        d = relentless.FixedKeyDict(keys=('A','B'))
+
+        #test copying for empty dict
+        dict_var = {'A':None, 'B':None}
+        self.assertEqual(d.todict(), dict_var)
+
+        #test copying for partially filled dict
+        dict_var = {'A':None, 'B':1.0}
+        d['B'] = 1.0
+        self.assertEqual(d.todict(), dict_var)
+
+        #test copying for full dict
+        dict_var = {'A':1.0, 'B':1.0}
+        d['A'] = 1.0
+        self.assertEqual(d.todict(), dict_var)
+
+if __name__ = '__main__':
+    unittest.main()
